adding paging to the things data

+++ trunk/Core/src/com/gawkat/core/client/account/thing/ThingsData.java
@@ -5,7 +5,7 @@ import com.google.gwt.user.client.rpc.IsSerializable;
 
 public class ThingsData implements IsSerializable {
 
-  public int total = 0;
+  public long total = 0;
   
   public ThingData[] thingData = null;
