+++ trunk/Core/src/com/gawkat/core/client/account/thingtype/ThingTypesData.java
@@ -4,7 +4,7 @@ import com.google.gwt.user.client.rpc.IsSerializable;
 
 public class ThingTypesData implements IsSerializable {
 
-  public int total = 0;
+  public long total = 0;
   
   public ThingTypeData[] thingTypeData = null;
